VueRouter: Move routes.ts file with app routing to constants from routes
